renamed PasswordProcessor to SecurityUtils and created GenerateRandomSalt method
